+++ RELEASE-NOTES.md
@@ -1,3 +1,5 @@
 # 1.0.3
 * Mavlink log files are now emitted to the logs directory (on Android versions this is in /sdcard/andropilot)
+* Serial port connection/disconnection should be reliable now
+